Fix for fastpack: Don't build if static libssl is present (#1)
* Figure out where binaries

* Ignore extension

* Check for existing libssl.a

* Remove unnecessary exit

* Fix bash script issue

+++ esy/build.sh
@@ -1,4 +1,13 @@
 cd _build
 
-make VERBOSE=1
-make install VERBOSE=1
+echo "Checking for existing SSL installs..."
+ls -a /lib/libssl*
+ls -a /usr/lib/libssl*
+
+if [ -e /usr/lib/libssl.a ]; then
+    echo "libssl.a already available; not building"
+else
+    make VERBOSE=1
+    make install VERBOSE=1
+fi
+
